pyproject.toml
[project]
name = "project-currency-ingestion"
version = "0.1.0"
description = ""
authors = [
    {name = "Durso",email = "dursoarthur@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "dlt (>=1.24.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
